[#200] move : 조회용 서비스의 경우 query 디렉토리로 이동
